// userManagmentBackend/src/ai/ai.types.ts
// this is like contract to the ai ,this thing i am telling any ai to understand it ,what exactly ai allowed to thats it 

export enum jarvisTypes {
  DEVICE_STATUS = "DEVICE_STATUS",

  DEVICE_ANALYTICS = "DEVICE_ANALYTICS",

  GATEWAY_STATUS = "GATEWAY_STATUS",

  SYSTEM_HEALTH = "SYSTEM_HEALTH",

  SCHEDULARS = "SCHEDULARS",

  REPORTS = "REPORTS",

  TELEMETRY = "TELEMETRY",

  PLATFORM_HELP = "PLATFORM_HELP",

  OUT_OF_SCOPE = "OUT_OF_SCOPE",

  DEVICE_TRIGGERING = "DEVICE_TRIGGERING"
}